feat(logs): logs to file with rotating + code split into files
